feat: refine API draft
